+++ timeRangeLib/timeRangeValidator.go
@@ -12,14 +12,6 @@ func (tr TimeRange) ValidateTimeRange() error {
 		return errors.New("invalid Frequency type")
 	}
 	if tr.Frequency != Fixed {
-		colonCountFrom := strings.Count(tr.HourMinuteFrom, ":")
-		if colonCountFrom != 1 {
-			return errors.New("invalid format: must contain exactly one colon")
-		}
-		colonCountTo := strings.Count(tr.HourMinuteTo, ":")
-		if colonCountTo != 1 {
-			return errors.New("invalid format: must contain exactly one colon")
-		}
 		err := validateHourMinute(tr.HourMinuteFrom)
 		if err != nil {
 			return err
@@ -42,7 +34,7 @@ func (tr TimeRange) ValidateTimeRange() error {
 		if tr.TimeFrom.After(tr.TimeTo) {
 			return errors.New("TimeFrom must be less than TimeTo for Fixed frequency")
 		}
-		if tr.TimeFrom == tr.TimeTo {
+		if tr.TimeFrom.Equal(tr.TimeTo) {
 			return errors.New("TimeFrom must not be equal to TimeTo for Fixed frequency")
 		}
 	case Weekly:
@@ -60,7 +52,7 @@ func (tr TimeRange) ValidateTimeRange() error {
 		if tr.DayFrom == 0 || tr.DayTo == 0 {
 			return errors.New("DayFrom, DayTo, must be present for Monthly frequency")
 		}
-		if tr.DayFrom == tr.DayTo && tr.compareHourMinute() {
+		if tr.DayFrom == tr.DayTo && isFromBeforeTo(tr.HourMinuteFrom, tr.HourMinuteTo) {
 			return errors.New("Invalid value of hourMinuteFrom or hourMinuteTo  for same day ,hourMinuteFrom >hourMinuteTo")
 		}
 		// this is to prevent overlapping windows crossing to next month for both negatives
@@ -83,19 +75,13 @@ func validateHourMinute(hourMinute string) error {
 		return errors.New("HourMinute is not valid, should be strictly of format HH:MM")
 	}
 	hh, err := strconv.Atoi(parts[0])
-	if err != nil {
-		return errors.New("Hour is not valid" + parts[0])
-	}
-	if hh > 23 || hh < 0 {
+	if err != nil || hh > 23 || hh < 0 {
 		return errors.New("Hour is not valid" + parts[0])
 	}
 
 	mm, err := strconv.Atoi(parts[1])
-	if err != nil {
-		return errors.New("Hour is not valid" + parts[1])
-	}
-	if mm > 59 || mm < 0 {
-		return errors.New("Hour is not valid" + parts[0])
+	if err != nil || mm > 59 || mm < 0 {
+		return errors.New("Minute is not valid" + parts[1])
 	}
 	return nil
 }
